Support creation of CloseContacts via API
