+++ app/helpers/validation_helper.rb
@@ -205,7 +205,8 @@ module ValidationHelper # rubocop:todo Metrics/ModuleLength
     report: { label: 'Lab Report Date', checks: [:date] },
     result: { label: 'Result', checks: [:enum] },
     assigned_user: { label: 'Assigned User', checks: [] },
-    continuous_exposure: { label: 'Continuous Exposure', checks: [:bool] }
+    continuous_exposure: { label: 'Continuous Exposure', checks: [:bool] },
+    patient_id: { label: 'Patient ID', checks: [] }
   }.freeze
 
   # Validates if a given date value is between (inclusive) two dates.
@@ -235,7 +236,7 @@ module ValidationHelper # rubocop:todo Metrics/ModuleLength
 
       # NOTE: If the value is a date, the typecast value may not correspond to original user input, so get value_before_type_cast
       value = VALIDATION[attribute][:checks].include?(:date) ? resource.public_send("#{attribute}_before_type_cast") : resource[attribute]
-      msg_header = (value ? " Value '#{value}' for " : '') + "'#{VALIDATION[attribute][:label]}'"
+      msg_header = (value ? "Value '#{value}' for " : '') + "'#{VALIDATION[attribute][:label]}'"
       errors.each do |error_message|
         # Exclude the actual value in logging to avoid PII/PHI
         Rails.logger.info "Validation Error on: #{attribute}"
